added extra checks.
git-svn-id: http://svn.php.net/repository/pear/packages/LiveUser_Admin/trunk@175067 c90b9560-bf6c-de11-be94-00142212c4b1

+++ Perm/Medium.php
@@ -97,9 +97,13 @@ class LiveUser_Admin_Perm_Medium extends LiveUser_Admin_Perm_Simple
     function removeGroup($filters)
     {
         // Remove users from the group
-        $result = $this->_storage->delete('groupusers', $filters);
-        if (!$result) {
-            return false;
+        $filter_check = array('group_id' => $filters['group_id']);
+        $count = $this->_storage->selectOne('groupusers', 'group_id', $filter_check, true);
+        if ($count > 0) {
+            $result = $this->_storage->delete('groupusers', $filters);
+            if (!$result) {
+                return false;
+            }
         }
 
         // Delete group rights
